Consistent path attribute for file-related modules
Not all file-related modules consistently use "path" as the attribute to
specify a single filename, some use "dest", others use "name". Most do
have aliases for either "name" or "destfile".

This change makes "path" the default attribute for (single) file-related
modules, but also adds "dest" and "name" as aliases, so that people can
use a consistent way of attributing paths, but also to ensure backward
compatibility with existing playbooks.

If this is acceptable, I don't mind doing the same for the extras
repository, and other modules.